Improved readme and renaming
