+++ Cargo.toml
@@ -9,7 +9,7 @@ members = [
     "hacspec-examples/hacspec-chacha20",
     "hacspec-examples/hacspec-poly1305",
     "hacspec-examples/hacspec-chacha20poly1305",
-    "spec-examples",
+    "unsafe-hacspec-examples",
     "hacspec-attributes",
     "hacspec-sig"
 ]
@@ -22,7 +22,7 @@ default-members = [
     "hacspec-examples/hacspec-chacha20",
     "hacspec-examples/hacspec-poly1305",
     "hacspec-examples/hacspec-chacha20poly1305",
-    "spec-examples",
+    "unsafe-hacspec-examples",
     "hacspec-attributes",
     "hacspec-sig"
 ]
