+++ unsafe-hacspec-examples/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "hacspec-examples"
+name = "unsafe-hacspec-examples"
 version = "0.1.0"
 authors = ["Franziskus Kiefer <franziskuskiefer@gmail.com>"]
 edition = "2018"
