+++ unsafe-hacspec-examples/src/hmac/mod.rs
