+++ unsafe-hacspec-examples/src/lib.rs
