+++ unsafe-hacspec-examples/src/ntru_prime/ntru_prime.rs
